move metrics to loadtest controller and delete from observability
