+++ cmd/controller.go
@@ -2,8 +2,13 @@ package cmd
 
 import (
 	"fmt"
+	"go.opentelemetry.io/otel/exporters/prometheus"
+	"log"
 	"strings"
 	"time"
+
+	"go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 
 	"github.com/kelseyhightower/envconfig"
 	"github.com/spf13/cobra"
@@ -16,6 +21,10 @@ import (
 	clientSet "github.com/hellofresh/kangal/pkg/kubernetes/generated/clientset/versioned"
 	informers "github.com/hellofresh/kangal/pkg/kubernetes/generated/informers/externalversions"
 )
+
+// reconcileDistribution defines the bucket boundaries for the histogram of reconcile latency metric
+// Bucket boundaries are 10ms, 100ms, 1s, 10s, 30s and 60s.
+var reconcileDistribution = []float64{10, 100, 1000, 10000, 30000, 60000}
 
 type controllerCmdOptions struct {
 	kubeConfig           string
@@ -51,7 +60,11 @@ func NewControllerCmd() *cobra.Command {
 				return fmt.Errorf("could not build logger instance: %w", err)
 			}
 
-			pe := observability.NewOtelPromExporter()
+			pe, err := prometheus.New()
+			if err != nil {
+				log.Fatal(err)
+				return nil
+			}
 
 			kubeCfg, err := kubernetes.BuildClientConfig(cfg.MasterURL, cfg.KubeConfig, cfg.KubeClientTimeout)
 			if err != nil {
@@ -68,7 +81,13 @@ func NewControllerCmd() *cobra.Command {
 				return fmt.Errorf("error building kangal clientSet: %w", err)
 			}
 
-			statsClient, err := observability.NewMetricReporter()
+			provider := metric.NewMeterProvider(metric.WithReader(pe), metric.WithView(metric.NewView(
+				metric.Instrument{Name: "kangal_reconcile_latency"},
+				metric.Stream{Aggregation: aggregation.ExplicitBucketHistogram{
+					Boundaries: reconcileDistribution,
+				}},
+			)))
+			statsReporter, err := controller.NewMetricsReporter(provider.Meter("controller"))
 			if err != nil {
 				return fmt.Errorf("error getting stats client:  %w", err)
 			}
@@ -81,7 +100,7 @@ func NewControllerCmd() *cobra.Command {
 				Exporter:       pe,
 				KubeClient:     kubeClient,
 				KangalClient:   kangalClient,
-				StatsReporter:  statsClient,
+				StatsReporter:  statsReporter,
 				KubeInformer:   kubeInformerFactory,
 				KangalInformer: kangalInformerFactory,
 			})
