+++ pkg/controller/loadtest.go
@@ -3,6 +3,11 @@ package controller
 import (
 	"context"
 	"fmt"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/metric/instrument"
+	"go.opentelemetry.io/otel/metric/instrument/syncint64"
+	"go.opentelemetry.io/otel/metric/unit"
 	"time"
 
 	"go.uber.org/zap"
@@ -22,7 +27,6 @@ import (
 	"k8s.io/client-go/util/workqueue"
 
 	"github.com/hellofresh/kangal/pkg/backends"
-	"github.com/hellofresh/kangal/pkg/core/observability"
 	loadTestV1 "github.com/hellofresh/kangal/pkg/kubernetes/apis/loadtest/v1"
 	clientSetV "github.com/hellofresh/kangal/pkg/kubernetes/generated/clientset/versioned"
 	sampleScheme "github.com/hellofresh/kangal/pkg/kubernetes/generated/clientset/versioned/scheme"
@@ -35,6 +39,75 @@ const (
 	falseString         = "false"
 	trueString          = "true"
 )
+
+// MetricsReporter used to interface with the metrics configurations
+type MetricsReporter struct {
+	countRunningLoadtests syncint64.UpDownCounter
+	workQueueDepthStat    syncint64.UpDownCounter
+	reconcileCountStat    syncint64.UpDownCounter
+	reconcileLatencyStat  syncint64.Histogram
+}
+
+// NewMetricsReporter contains loadtest metrics definition
+func NewMetricsReporter(meter metric.Meter) (*MetricsReporter, error) {
+	countRunningLoadtests, err := meter.SyncInt64().UpDownCounter(
+		"kangal_running_loadtests_count",
+		instrument.WithDescription("The number of currently running loadtests"),
+		instrument.WithUnit(unit.Dimensionless),
+	)
+	if err != nil {
+		fmt.Errorf("could not register countRunningLoadtests metric: %w", err)
+		return nil, err
+	}
+
+	//if err := meter.RegisterCallback(
+	//	[]instrument.Asynchronous{
+	//		countRunningLoadtests,
+	//	},
+	//	func(ctx context.Context) {
+	//		countRunningLoadtests.Observe(context.Background(), 1, attribute.String("loadtest", "running"))
+	//	},
+	//); err != nil {
+	//	panic(err)
+	//}
+
+	workQueueDepthStat, err := meter.SyncInt64().UpDownCounter(
+		"kangal_work_queue_depth",
+		instrument.WithDescription("Depth of the work queue"),
+		instrument.WithUnit(unit.Dimensionless),
+	)
+	if err != nil {
+		fmt.Errorf("could not register workQueueDepthStat metric: %w", err)
+		return nil, err
+	}
+
+	reconcileCountStat, err := meter.SyncInt64().UpDownCounter(
+		"kangal_reconcile_count",
+		instrument.WithDescription("Number of reconcile operations"),
+		instrument.WithUnit(unit.Dimensionless),
+	)
+	if err != nil {
+		fmt.Errorf("could not register reconcileCountStat metric: %w", err)
+		return nil, err
+	}
+
+	reconcileLatencyStat, err := meter.SyncInt64().Histogram(
+		"kangal_reconcile_latency",
+		instrument.WithDescription("Latency of reconcile operations"),
+		instrument.WithUnit(unit.Milliseconds),
+	)
+	if err != nil {
+		fmt.Errorf("could not register reconcileLatencyStat metric: %w", err)
+		return nil, err
+	}
+
+	return &MetricsReporter{
+		countRunningLoadtests: countRunningLoadtests,
+		workQueueDepthStat:    workQueueDepthStat,
+		reconcileCountStat:    reconcileCountStat,
+		reconcileLatencyStat:  reconcileLatencyStat,
+	}, nil
+}
 
 // Controller is the controller implementation for LoadTest resources
 type Controller struct {
@@ -61,7 +134,7 @@ type Controller struct {
 	// Kubernetes API.
 	recorder record.EventRecorder
 
-	statsClient observability.MetricsReporter
+	statsClient MetricsReporter
 
 	registry backends.Registry
 	logger   *zap.Logger
@@ -74,7 +147,7 @@ func NewController(
 	kangalClientSet clientSetV.Interface,
 	kubeInformerFactory informers.SharedInformerFactory,
 	kangalInformerFactory externalversions.SharedInformerFactory,
-	statsClient observability.MetricsReporter,
+	statsClient MetricsReporter,
 	registry backends.Registry,
 	logger *zap.Logger) *Controller {
 
@@ -208,7 +281,7 @@ func (c *Controller) processNextWorkItem() bool {
 	}
 
 	// Send the metrics for the current queue depth
-	//c.statsClient.ReportQueueDepth(int64(c.workQueue.Len()))
+	c.statsClient.workQueueDepthStat.Add(context.Background(), int64(c.workQueue.Len()))
 
 	// We wrap this block in a func, so we can defer c.workQueue.Done.
 	err := func(obj interface{}) error {
@@ -230,7 +303,9 @@ func (c *Controller) processNextWorkItem() bool {
 			if err != nil {
 				status = falseString
 			}
-			c.statsClient.ReportReconcile(time.Now().Sub(startTime), key, status)
+
+			c.statsClient.reconcileCountStat.Add(context.Background(), 1, attribute.String("key", key), attribute.String("success", status))
+			c.statsClient.reconcileLatencyStat.Record(context.Background(), int64(time.Since(startTime)/time.Millisecond), attribute.String("key", key), attribute.String("success", status))
 		}()
 
 		// We expect strings to come off the workQueue. These are of the
@@ -276,7 +351,7 @@ func (c *Controller) syncHandler(key string) error {
 	ctx, cancel := context.WithTimeout(context.Background(), c.cfg.SyncHandlerTimeout)
 	defer cancel()
 
-	c.statsClient.AddRunningLTCounter(c.countRunningLoadtests())
+	c.statsClient.countRunningLoadtests.Add(context.Background(), c.countRunningLoadtests(), attribute.String("loadtest", "running"))
 
 	logger := c.logger.With(
 		zap.String("loadtest", key),
@@ -384,7 +459,7 @@ func (c *Controller) handleObject(obj interface{}) {
 		}
 
 		c.logger.Debug("Processing object", zap.String("object-name", object.GetName()))
-		c.statsClient.AddRunningLTCounter(c.countRunningLoadtests())
+		c.statsClient.countRunningLoadtests.Add(context.Background(), c.countRunningLoadtests(), attribute.String("loadtest", "running"))
 
 		foo, err := c.loadtestsLister.Get(ownerRef.Name)
 		if err != nil {
